Add the result files to cargo-build_target BYPRODUCTS
